checkpoint, fixing via processing
